Implemented form definition template overrides
In certain instances it is useful to be able to use a custom template but still maintain all the standard behavior of a type : specifically I was implementing bootstrap's form-horizontal for certain forms, but didn't want to have to completely replace all the templates. By adding this templateOverride feature, I can copy the default decorator's templates for a specific type (say, array) and then add the appropriate extra markup. I can then just set the templateOverride field and then it will use the new template while maintaining all the existing functionality.

This makes small template tweaks really easy without using the template type (which loses the behavior of the underlying array/fieldset). It does require a certain amount of expertise and knowledge about the templates but has permitted me to spot-customize the templates for different forms (which is painful to do as I don't believe there is currently an easy way to load multiple decorators and pick which one gets used on a specific schema entry) as well as use alternative sorting implementations (I didn't want to add all of jQuery UI just for ui-sortable, and I was able to use a smaller sortable implementation).